requirements.txt
opencv-python
#opencv-python-headless
numpy
pyserial
tflite-runtime
ultralytics
#ultralytics-opencv-headless